Move learning_curve module into model_selection.validate
